requirement.txt
Django==3.1.4
djangorestframework==3.13.1
django-cors-headers==3.13.0
django-environ==0.9.0
moviepy==1.0.3
